backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Login
Werkzeug
python-dotenv
gunicorn
psycopg2-binary
